setting up project
