+++ .github/release-drafter.yml
@@ -4,19 +4,6 @@ categories:
   - title: 'Dependencies'
     labels:
       - 'dependencies'
-
-# todo - is this needed, can just delete and not have labels?
-version-resolver:
-  major:
-    labels:
-      - 'major'
-  minor:
-    labels:
-      - 'minor'
-  patch:
-    labels:
-      - 'patch'
-  default: minor
 template: |
   # Changes
