+++ README.md
@@ -1,14 +1,28 @@
 # raspimonitor
 Monitoring and alerting for your RaspberryPi
 
-# todo - check make areadme
+## Installation
 
-# todo - pin in github once ready
-# obsidian note?
+TODO
 
-# todo - add inspiration form
-https://github.com/git-ogawa/raspi-statmon/blob/main/src/rstatmon/statdata.py
 
-https://github.com/mityax/rpi-temperature-alert/tree/master
+## Usage
 
-https://www.howtogeek.com/discord-slack-alert-raspberry-pi-too-hot/
+TODO
+
+## Contributing
+
+Pull requests are welcome. 
+
+Please make sure to update tests as appropriate.
+
+## License
+
+[MIT](https://choosealicense.com/licenses/mit/)
+
+[License file](LICENSE)
+
+## Obtained inspiration from
+- https://github.com/git-ogawa/raspi-statmon/blob/main/src/rstatmon/statdata.py
+- https://github.com/mityax/rpi-temperature-alert/tree/master
+- https://www.howtogeek.com/discord-slack-alert-raspberry-pi-too-hot/
