+++ requirements-dev.txt
@@ -1,0 +1,45 @@
+#
+# This file is autogenerated by pip-compile with Python 3.9
+# by the following command:
+#
+#    make freeze-dev
+#
+build==1.1.1
+    # via pip-tools
+certifi==2024.2.2
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+click==8.1.7
+    # via pip-tools
+idna==3.6
+    # via requests
+importlib-metadata==7.0.2
+    # via build
+packaging==23.2
+    # via build
+pip-tools==7.4.1
+    # via raspimonitor (pyproject.toml)
+pyproject-hooks==1.0.0
+    # via
+    #   build
+    #   pip-tools
+requests==2.31.0
+    # via raspimonitor (pyproject.toml)
+ruff==0.3.1
+    # via raspimonitor (pyproject.toml)
+tomli==2.0.1
+    # via
+    #   build
+    #   pip-tools
+    #   pyproject-hooks
+urllib3==2.2.1
+    # via requests
+wheel==0.42.0
+    # via pip-tools
+zipp==3.17.0
+    # via importlib-metadata
+
+# The following packages are considered to be unsafe in a requirements file:
+# pip
+# setuptools
