+++ requirements.txt
@@ -1,1 +1,16 @@
+#
+# This file is autogenerated by pip-compile with Python 3.9
+# by the following command:
+#
+#    make freeze
+#
+certifi==2024.2.2
+    # via requests
+charset-normalizer==3.3.2
+    # via requests
+idna==3.6
+    # via requests
 requests==2.31.0
+    # via raspimonitor (pyproject.toml)
+urllib3==2.2.1
+    # via requests
